requirements.txt
pillow
pytest
pytest-mock
pytest-coverage
